ARM: tegra: Set NVC config for cardhu-focusers
* NVC framework provides NVC_CFG_NODEV config for all
NVC compliant camera drivers to make sure to delete
the device nodes which are not found in I2C probing.
This config is not set for all focuser modules on PM269,
which introduces the issue.

* Earlier only sh532u focuser was suppose to be on
cardhu, hence this error condition did not get triggered
but newly added focuser has exposed this error.

* Set NVC_CFG_NODEV for both sh532u and ad5816 focusers.

* Add regulator supply for cam2 ad5816 focuser, it fixes
camera crash on first reboot on E1198 cardhu and I2C
warnings.

Bug 992611
Bug 995072

Change-Id: I3382a896e4e246a0a323bd13ae0889b6701351c7
Signed-off-by: Sudhir vyas <svyas@nvidia.com>
Reviewed-on: http://git-master/r/106960
Reviewed-by: Automatic_Commit_Validation_User
Reviewed-by: Prayas Mohanty <pmohanty@nvidia.com>
Reviewed-by: Sachin Nikam <snikam@nvidia.com>
Reviewed-by: Frank Chen <frankc@nvidia.com>