video: tegra: nvavp: add support for vde.emc
vde.emc is required to know whether video playback is
going on or not. Display efficiency rate is set to 45%
if both dc.emc and vde.emc is enabled.

vde.emc clock is enabled when non-zero avp.emc clock
rate is requested and is disabled when zero avp.emc clock
is requested.

Bug 1266369

Change-Id: Ic1de7145763a41a8adf11a9e1cd7a1cd723f46e2
Signed-off-by: Kevin Huang <kevinh@nvidia.com>
Reviewed-on: http://git-master/r/232614
Reviewed-by: Soumenkumar Dey <sdey@nvidia.com>
Tested-by: Soumenkumar Dey <sdey@nvidia.com>
Reviewed-by: Bharat Nihalani <bnihalani@nvidia.com>