ARM: tegra: tegratab: memory: correct MC_GRANT_DECREMENT value
This is due to an incorrect value for GRANT DECREMENT in the DVFS table for
frequencies less than 400 MHz (EMC clock). On T114, we are always using
MC:EMC freq of 1:2. However, the code for calculating GRANT DECREMENT in the
DVFS table logic assumes MC:EMC freq of 1:1 at EMC freq < 400. Therefore,
the value in the DVFS table at EMC < 400 is double what it should be.

Bug 1299452
Bug 1277921

Change-Id: I5de80bc1b46e425e372cc5b461e425060553a3c3
Signed-off-by: Min-wuk Lee <mlee@nvidia.com>
Signed-off-by: Jay Cheng <jacheng@nvidia.com>
Reviewed-on: http://git-master/r/236614
Reviewed-by: Automatic_Commit_Validation_User
Reviewed-by: Bharat Nihalani <bnihalani@nvidia.com>