ARM: tegra11: clock: Add cbus fine granularity region
Added fine granularity region to cbus possible rates. In this region
requested cbus rate is not clipped to dvfs steps, but rounded to fine
granularity resolution. The latter is set as 12MHz, and the region is
defined as 5 resolution steps below the top dvfs rate, assuming this
top rate is reachable on the particular chip bin/sku.

Change-Id: Id111b074c39a0a314d57da8ce628f3dbbf385b73
Signed-off-by: Alex Frid <afrid@nvidia.com>
Reviewed-on: http://git-master/r/249902
Reviewed-by: Automatic_Commit_Validation_User
Reviewed-by: Yu-Huan Hsu <yhsu@nvidia.com>