arm: tegra: macallan: memory: correct MC_GRANT_DECREMENT value
This is due to an incorrect value for GRANT DECREMENT in the DVFS table for
frequencies less than 400 MHz (EMC clock). On T114, we are always using
MC:EMC freq of 1:2. However, the code for calculating GRANT DECREMENT in the
DVFS table logic assumes MC:EMC freq of 1:1 at EMC freq < 400. Therefore,
the value in the DVFS table at EMC < 400 is double what it should be.

bug 1321312

Change-Id: I2b1eb7824077e621d8fdf9dd299a7a6013974b2a
Signed-off-by: Hayden Du <haydend@nvidia.com>
Reviewed-on: http://git-master/r/252125
GVS: Gerrit_Virtual_Submit
Reviewed-by: Daniel Fu <danifu@nvidia.com>
Tested-by: Daniel Fu <danifu@nvidia.com>
Reviewed-by: Yu-Huan Hsu <yhsu@nvidia.com>