arm: tegra: add FIQ WAR for t11x non-secure
On Tegra T11x WDT FIQ is not routed to the GIC, so we cannot use Tegra
WDT for soft-hang debugging.

This WAR is to configure GIC group0 to generate the FIQ to CPU, and only
let WDT routed to GIC group0. All other normal IRQs will use GIC group1.
This WAR is only applicable to non-secure OS build; secure OS build will
treat FIQ as secure interrupt for enterring monitor.

This patch is based on previous work of Lucas Dai and Bo Yan.

bug 1277869
bug 1357562

Change-Id: Ie9a05f9f2974d2b9fb2d09f7847639e553917cb7
Signed-off-by: Xin Xie <xxie@nvidia.com>
Reviewed-on: http://git-master/r/268797
Reviewed-by: Mrutyunjay Sawant <msawant@nvidia.com>
Tested-by: Mrutyunjay Sawant <msawant@nvidia.com>