arm:tegra: Correct sample rate for voice call
Because the channel1 of DAM is used for DL, there is no SRC in DL.
the  codec need to be set the same rate with baseband.
Otherwise the I2S Timing REG cannot get the suited value and
left channel may be swapped with right channel.

Bug 1441449

Change-Id: Ib71bb41a7e03d2a85d718be0c94a5a5f525c0ce3
Signed-off-by: Lei Fan <leif@nvidia.com>
Reviewed-on: http://git-master/r/358078
Reviewed-by: Automatic_Commit_Validation_User
Reviewed-by: Vijay Mali <vmali@nvidia.com>
GVS: Gerrit_Virtual_Submit
Reviewed-by: Martin Chi <mchi@nvidia.com>